+++ word_embeddings/word2vecfw/word2vecf_lib.h
@@ -16,24 +16,24 @@ typedef float real;                    // Precision of float numbers
 
 #define N 40                  // number of closest words that will be shown
 
-long long GetFileSize(char *fname);
+int64_t GetFileSize(char *fname);
 
 
 // Used for sampling of negative examples.
 // wc[i] == the count of context number i
 // wclen is the number of entries in wc (context vocab size)
-void InitUnigramTable(struct vocabulary *v, int **unitable, int table_size);
+void InitUnigramTable(struct vocabulary *v, int32_t **unitable, int32_t table_size);
 
-void InitNet(struct vocabulary *wv, struct vocabulary *cv, real **word_vecs, real **context_vecs, long long layer1_size);
+void InitNet(struct vocabulary *wv, struct vocabulary *cv, real **word_vecs, real **context_vecs, int64_t layer1_size);
 
 void InitRelationWeight(struct vocabulary *rv, real **relation_weight);
 
-void dumpVec(char *dump_file, struct vocabulary *v, real *vecs, long long layer1_size, int binary);
+void dumpVec(char *dump_file, struct vocabulary *v, real *vecs, int64_t layer1_size, int32_t binary);
 
-void dumpKMean(long long classes, char *dump_file, struct vocabulary *wv, real *word_vecs, long long layer1_size);
+void dumpKMean(int64_t classes, char *dump_file, struct vocabulary *wv, real *word_vecs, int64_t layer1_size);
 
-int loadVec(char *fname, long long *words, long long *size, char **vocab, float **vecs, int normalize);
+int32_t loadVec(char *fname, int64_t *words, int64_t *size, char **vocab, float **vecs, int32_t normalize);
 
 void grab_relation(char *relation, char *word);
 
-int ArgPos(char *str, int argc, char **argv);
+int32_t ArgPos(char *str, int32_t argc, char **argv);
